+++ plugins/passkey/CHANGELOG.md
@@ -1,25 +1,37 @@
 # @zerodev/passkey-validator
 
+## 5.3.0
+
+### Minor Changes
+
+- Kernel v3.1 released. And kernel versioning update. Added kernelVersion param in the interface.
+
+## 5.3.0-alpha-0
+
+### Minor Changes
+
+- Integrated Kernel v3.1 and added kernelVersion param in account and plugins interface
+
 ## 5.2.8
 
 ### Patch Changes
 
--   Decoupled creating passkey itself and creating the validator
+- Decoupled creating passkey itself and creating the validator
 
 ## 5.2.5
 
 ### Patch Changes
 
--   Remove usePrecompiled from enable data
+- Remove usePrecompiled from enable data
 
 ## 5.2.4
 
 ### Patch Changes
 
--   Fix Entrypoint version in getPasskeyValidator
+- Fix Entrypoint version in getPasskeyValidator
 
 ## 5.2.3
 
 ### Patch Changes
 
--   Update permissionless@0.1.18
+- Update permissionless@0.1.18