+++ plugins/social/CHANGELOG.md
@@ -1,1 +1,13 @@
 # @zerodev/social-validator
+
+## 5.1.0
+
+### Minor Changes
+
+- Kernel v3.1 released. And kernel versioning update. Added kernelVersion param in the interface.
+
+## 5.1.0-alpha-0
+
+### Minor Changes
+
+- Integrated Kernel v3.1 and added kernelVersion param in account and plugins interface